Add containers for project examples
Created a section-0 container for the first flex-box section after
the header-bar. Within section-0 added an examples container for two
containers. Plan is to have a structure of section-container,
example-container, and two info-containers per section. The pictures
will have anchor tags linking to a page with more photos related
to the corresponding linked photo on the project page.