Style order list style types and indentation, style sidebar selectors
Styled ordered lists nested within other ordered lists to be indented
i.e. set margin-left property. Additionally, styled the position and
colors of the selectors. Idea is to highlight a selected section
by changing its CSS. Next step is to style only the selected section and
to set up style for selectors that are hovered over.