Style selector:hover, active, and .sidebar .selector
Added same style to selector:hover and active to display the
hovered-over or selected selector in the sidebar with background
and font color changes. The active class is used to differentiate the
selected element from the non-selected elements in the sidebar. The
JS script adds the active class to the selected element and removes
it from the sibling elements.