Refactor and comment
Added an eventHandler() function and a displaySelectedElement()
function. The displaySelectedElement() function sets display to
'block' for the corresponding selected element and sets display to none
for sibling elements. Similarly, the displaySelectedElement() function
adds the active class to the selected element and removes it from
sibling elements. Deleted reference to content section in favor
of dynamic reference of current selected element within for
displaySelectedElement() function.