Create container with class menu inside sidebar
The container within sidebar allows the elements within the container
to display in block format i.e. vertically which avoids need to set the
flex-direction on the sidebar to column. This is important because it
allows to set background color of the sidebar and min-height to 100vh
in order to allow background color to display full vertical length of
the screen.