Remove display flex from .sidebar
Added menu container within sidebar which allows the elements within
the container to display in block format i.e. vertically which avoids
need to set the flex-direction on the sidebar to column. This is
important because it allows background color of the sidebar to be set
together with setting min-height to 100vh in order to allow background
color to display full ßvertical length of the screen.